==> 1_ETL_ITEMS.ipynb <==
## Importamos las librerias

import pandas as pd
from pandas import json_normalize
import numpy as np
import json
import warnings
warnings.filterwarnings("ignore")
import ast

## Cargamos el archivo JSON, se crea en un DataFrame y se analiza


# Se carga el archivo JSON
ruta_items = "data_repo/australian_users_items.json"

# se crea una lista para almacenar las filas del archivo JSON
filas_items = []

# Se abre el archivo JSON en modo lectura, asegurándose de usar la codificación UTF-8
with open(ruta_items, encoding="utf-8") as f:
    # se itera sobre cada línea del archivo
    for line in f.readlines():
        # se convierte la línea en un diccionario Python utilizando ast.literal_eval()
        # Esto es seguro ya que sabemos que las líneas contienen expresiones literales de Python válidas (diccionarios)
        fila_dict = ast.literal_eval(line)
        # se agrega el diccionario a la lista de filas
        filas_items.append(fila_dict)

# se crea un DataFrame de Pandas a partir de la lista de diccionarios
df_items = pd.DataFrame(filas_items)
df_items

df_items.columns

df_items.info()

Se observa que los tipos de datos son objetos y la cantidad de valores no-nulos

Se visualiza que se encuentran los juegos por cada usuario con las siguientes variables:
* user_id : identificador unico de usuario
* user_url : URL perfil del usuario
* items : Items de usuario en formato Json

Trabajos a realizar : 
* Limpieza de valores nulos y duplicados
* Trabajar las variables anidadas
* Borrar las columnas que no son importantes para el analisis


## Trabajaremos con la variable items

la variable items se encuentra con datos anidados, se realiza el trabajo para desnadiar 

# se realiza con el método explode() para expandir la columna "items"
df_items = df_items.explode("items")

# se reinicia el índice del DataFrame
df_items = df_items.reset_index(drop=True)

# se muestra el DataFrame resultante
df_items

# Se normaliza la columna "items" del DataFrame df_items
df_items_normalizado = pd.json_normalize(df_items["items"])

# se muestra el DataFrame resultante
df_items_normalizado

# Se concatena los DataFrames df_items y df_items_normalizado a lo largo de las columnas
df_items_final = pd.concat([df_items, df_items_normalizado], axis=1)

# Se muestra el DataFrame resultante
df_items_final

# Se elimina la columna "items" del DataFrame df_items_final
df_items_final = df_items_final.drop(columns=["items"])
df_items_final

#Se elimina las columnas que no usaremos
df_items_final = df_items_final.drop(columns=["user_url","playtime_2weeks","steam_id"])
df_items_final

df_items_final["item_name"].unique()

# Se define el patrón no deseado
patron_no_deseado = r'[^a-zA-Z0-9\s]'

# Se filtra el DataFrame df_items_final
# Seleccionar las filas donde el valor de la columna "item_name" no contiene el patrón no deseado
df_filtrado = df_items_final[df_items_final["item_name"].str.contains(patron_no_deseado) == False]


df_filtrado


## Verificamos los datos Nulos, duplicados y se limpian

df_filtrado.isnull().sum()

duplicados = df_filtrado.duplicated().sum()
duplicados

# Se crea una máscara booleana para identificar las filas duplicadas en df_filtrado
duplicados_mask = df_filtrado.duplicated(keep=False)

# Se seleciona las filas duplicadas utilizando la máscara
filas_duplicadas = df_filtrado[duplicados_mask]
filas_duplicadas

se observa los duplicados y corresponden al mismo usuario pero con algunas variables iguales (user_id, utems_count, steam_id) pero varia el juego por lo tanto se decide conservarlas.

Se agrega una nueva columna playtime_hour, para ver la cantidad de horas jugadas

df_filtrado["playtime_hour"]= df_filtrado["playtime_forever"]/60
df_filtrado

El dataframe limpio de Items (Informacion de juegos por usuario) contiene las siguientes columnas:
* user_id : Id de usuario
* items_count : Cantidad de juegos por usuario
* item_id : Id del juego
* item_name : Nombre de juego
* playtime_forever : Tiempo jugado en minutos
* playtime_hour : Timpo juego en horas

archivo_limpio = "data_repo/users_items_limpio.csv"
df_filtrado.to_csv(archivo_limpio, index=False, encoding="utf-8")
print(f"Se guardó el archivo {archivo_limpio}")
