==> 1_ETL_OUTPUT.ipynb <==
## Importamos las librerias

#Se importa pandas y numpy para manejar los datos y también la biblioteca json para trabajar con archivos JSON.
import pandas as pd
import numpy as np
import json
import warnings
#se utiliza warnings.filterwarnings("ignore") para ignorar las advertencias durante la ejecución del código.
warnings.filterwarnings("ignore")

## Cargamos el archivo JSON, se crea en un DataFrame y se analiza

#Se cargan los datos. 
ruta_output = 'data_repo\output_steam_games.json'

#se procede a leer linea por linea para extraer la informacion dado que hay datos adicionales en el archivo JSON
with open(ruta_output,'r') as archivo:
    datos_output = [json.loads(line) for line in archivo]

#Se crea un DataFrame de pandas llamado df_output que contiene la información del archivo JSON
df_output = pd.DataFrame(datos_output)
df_output

df_output.columns

df_output.info()

Se observa que todos los tipos de datos son objetos y que hay una gran cantidad de valores no-nulos

Se visualiza que se encuentran la informacion de los juegos con las siguientes variables:
* publisher = Empresa publicadora del contenido
* genres = Genero del contenido (anidado)
* app_name = Nombre del contenido
* title = Titulo del contenido
* url = URL de publicación del contenido
* release_date = Fecha de lanzamiento
* tags = etiquetas de contenido (anidado)
* reviews_url = Reviews de contenido
* specs = Especificaciones (anidado)
* price = Precio del contenido
* early_access = acceso temprano
* id = identificador unico de contenido
* developer = Desarrollador

Trabajos a realizar : 
* Limpieza de valores nulos y duplicados
* Trabajar las variables anidadas
* Borrar las columnas que no son importantes para el analisis


## Verificamos los datos Nulos y se limpian

df_output.isnull().sum()

Eliminamos filas que contienen datos nulos

#Al especificar how="all", estás indicando que solo se eliminan las filas donde todos los valores son nulos.
#Al agregar reset_index(drop=True), se reinicia los índices del DataFrame después de eliminar las filas nulas y descartando los índices antiguos.
df_output = df_output.dropna(how="all").reset_index(drop=True)

df_output.isnull().sum()

Eliminamos valores nulos de la variable id

df_output.dropna(subset=['id'], inplace=True)

df_output.isnull().sum()

## Verificamos filas duplicadas y se eliminan

df_output[df_output["id"].duplicated()]

#El argumento keep='first' hace que se mantenga la primera de las filas duplicadas y elimine la siguiente.
df_output = df_output.drop_duplicates(subset='id', keep='first')

df_output.isnull().sum()

## Trabajaremos con la Variable Price

df_output["price"].unique()

Se observa que hay 2 tipos de datos, que son str y float, para poder trabajar se necesita que los valores queden como float, se observa que las filas con valores que estan en str, todos indican que el juego fue gratis, por lo que se asignara un precio de 0.

# Se convierte la columna "price" en tipo numérico
# Se utiliza el parámetro errors="coerce", cualquier valor no convertible se convertirá en NaN (no un número)
df_output["price_numeric"] = pd.to_numeric(df_output["price"], errors="coerce")

# Se convierte la columna "price_numeric" en tipo numérico nuevamente para manejar cualquier valor no convertible restante
df_output["price_numeric"] = pd.to_numeric(df_output["price_numeric"], errors="coerce")

# Se reemplaza los valores NaN con 0
df_output["price_numeric"] = df_output["price_numeric"].fillna(0)

df_output["price_numeric"].dtype

## Trabajaremos con la variable release_date

df_output["release_date"].unique()

Se identifica las fechas que no siguen el formato "YYYY-MM-DD" y se almacena en un nuevo DataFrame llamado df_invalid_dates

# Se seleciona las fechas de lanzamiento que no siguen el formato "YYYY-MM-DD" utilizando una expresión regular
# Las fechas de lanzamiento que no siguen este formato se consideran inválidas
invalid_dates = df_output["release_date"][~df_output["release_date"].str.match(r"\d{4}-\d{2}-\d{2}", na=False, case=False, flags=0)]

# Se crea un nuevo DataFrame llamado df_invalid_dates que contiene las fechas de lanzamiento inválidas
df_invalid_dates = pd.DataFrame({"invalid_dates": invalid_dates})

# Se muestra el DataFrame con las fechas de lanzamiento inválidas
df_invalid_dates

Se cambia  el formato a la columna release_date y luego se crea una nueva columna release_year que es la variable que nos servira para los analisis.

# Se convierte la columna "release_date" al formato de fecha datetime
# La opción errors="coerce" convierte los valores no válidos en NaT (Not a Time)
# La opción format='%Y-%m-%d' especifica el formato de fecha esperado
# La opción exact=False permite que el formato sea flexible para aceptar diferentes representaciones de fecha
df_output["release_date"] = pd.to_datetime(df_output["release_date"], errors="coerce", format='%Y-%m-%d', exact=False)

# Se calcula la fecha media de lanzamiento
mean_release_date = df_output["release_date"].mean()

# Se llena los valores de fecha de lanzamiento que son NaT con la fecha media
df_output["release_date_filled"] = df_output["release_date"].fillna(mean_release_date)

# Se extrae el año de la fecha de lanzamiento y almacenarlo en una nueva columna "release_year"
df_output["release_year"] = df_output["release_date_filled"].dt.year

# Se muestra el DataFrame con las modificaciones realizadas
df_output.head(5)

df_output["release_year"].dtype

df_output["release_year"].unique()

df_output["release_year"].isnull().sum()

## Eliminaremos las columnas que no se usaran en los analisis 

# Se elimina las columnas especificadas del DataFrame df_output
# El parámetro axis=1 indica que se eliminan columnas en lugar de filas
# El parámetro inplace=True indica que los cambios se realizarán directamente en el DataFrame original
df_output.drop(["release_date", "price", "url", "publisher", "reviews_url", "specs", "early_access", "release_date_filled"], axis=1, inplace=True)

# Se muestra el DataFrame resultante después de eliminar las columnas
df_output.head(5)

## Trabajaremos con la variable Genres

se verifica si hay valores en "tags" y se seleciona aquellos que son géneros. Si encuentra géneros en "tags", los imputa en lugar del valor nulo en "genres" 

# Se crea una lista vacía para almacenar los géneros únicos
generos = []

# Se itera sobre cada lista de géneros en la columna "genres" del DataFrame
for genres_list in df_output["genres"].dropna():
    # Se itera sobre cada género en la lista de géneros
    for genre in genres_list:
        # se verifica si el género no está en la lista de géneros
        if genre not in generos:
            # Si el género no está en la lista, lo agregamos
            generos.append(genre)

# Se muestra la lista de géneros únicos
generos

# Se elimina los valores nulos de la columna "tags" y filtrar los elementos que pertenecen a los géneros
df_output["tags"] = df_output["tags"].dropna().apply(lambda lista: [g for g in lista if g in generos])

df_output["genres"].fillna(df_output["tags"], inplace=True)

df_output["genres"].isnull().sum()

se verifica que todavia hay registros nulos, que se eliminaran mas adelante
se elimina la columna tags 

df_output = df_output.drop("tags", axis=1)
df_output

# Se muestra las primeras filas de la columna "genres" en el DataFrame
df_output["genres"].head()

la variable genres se encuentra con datos anidados, se realiza el trabajo para desnadiar 

# Se agrega una nueva columna "main_genre" al DataFrame 
# La columna "main_genre" contendrá el primer género de la lista en la columna "genres"
# Si la columna "genres" está vacía o no es una lista, se asigna el valor None
df_output["main_genre"] = df_output["genres"].apply(lambda x: x[0] if isinstance(x, list) and x else None)

# Se elimina filas duplicadas en el DataFrame df_output basadas en varias columnas
# mantener solo la primera ocurrencia de cada conjunto único de valores en las columnas especificadas
df_unique_genre = df_output.drop_duplicates(["app_name", "title", "id", "developer", "price_numeric", "release_year"])

# Se elimina la columna "genres" del DataFrame df_unique_genre
# El parámetro axis=1 indica que se eliminará una columna en lugar de una fila
df_unique_genre = df_unique_genre.drop("genres", axis=1)

# Mostrar el DataFrame resultante df_unique_genre
df_unique_genre

df_unique_genre["main_genre"].isnull().sum()

# se llena los valores nulos en la columna "main_genre" con el valor "Sin Dato"
df_unique_genre["main_genre"] = df_unique_genre["main_genre"].fillna("Sin Dato")

# Se verifica cuántos valores nulos quedan en la columna "main_genre"
df_unique_genre["main_genre"].isnull().sum()

## Trabajaremos con las variables app_name, title y developer

# Lista de columnas a llenar
columnas = ["app_name", "title"]

# Se llena los valores nulos en las columnas especificadas con el valor "Sin datos"
df_unique_genre[columnas] = df_unique_genre[columnas].fillna("Sin datos")

# Se verifica si las columnas 'app_name' y 'title' son iguales en todos los registros
son_iguales = (df_unique_genre["app_name"] == df_unique_genre["title"]).all()

# Se imprime el resultado de la verificación
if son_iguales:
    print("Las columnas 'app_name' y 'title' son iguales en todos los registros.")
else:
    print("Las columnas 'app_name' y 'title' no son iguales en todos los registros.")


# Se agrega una nueva columna "son_iguales" al DataFrame df_unique_genre
# La columna contiene valores booleanos que indican si "app_name" y "title" son iguales en cada fila
df_unique_genre["son_iguales"] = df_unique_genre["app_name"] == df_unique_genre["title"]

# Se seleciona las filas donde la columna "son_iguales" es False
filas_no_iguales = df_unique_genre[df_unique_genre["son_iguales"] == False]

# Mostrar las filas donde "app_name" y "title" son diferentes
filas_no_iguales

# Se elimina las filas con valores nulos en la columna "developer" del DataFrame df_unique_genre
# El parámetro subset=["developer"] indica que solo se eliminarán las filas con valores nulos en la columna "developer"
df_unique_genre.dropna(subset=["developer"], inplace=True)

df_unique_genre.isnull().sum()

# Se elimina la columna "title" del DataFrame df_unique_genre
# El parámetro columns=["title"] indica que se eliminará la columna "title"
# El parámetro axis=1 indica que se eliminará una columna en lugar de una fila
df_unique_genre.drop(columns=["title"], axis=1, inplace=True)

df_unique_genre

df_unique_genre.drop(["son_iguales"], axis=1, inplace=True)

df_unique_genre.head(2)

df_unique_genre.isnull().sum()

El dataframe limpio de output steam games (Informacion de juegos) contiene las siguientes columnas
* app_name : Nombre del juego 
* id : Id de juego
* developer : desarrollador
* price_numeric : Precio de lanzamiento
* release_year : Año de lanzamiento
* main_genre : Genero principal

## Se cargara el dataset limpio

# se especifica el nombre del archivo donde se guardará el DataFrame
archivo_limpio = 'data_repo/output_steam_games_limpio.csv'

# se guarda el DataFrame en un archivo CSV
# El parámetro index=False evita que se agregue una columna de índice al archivo CSV
# El parámetro encoding='utf-8' especifica la codificación del archivo CSV
df_unique_genre.to_csv(archivo_limpio, index=False, encoding='utf-8')

# Se imprime un mensaje indicando que el archivo se guardó correctamente
print(f'Se guardó el archivo {archivo_limpio}')
