==> 3_DATAFRAME_FUNCIONES.ipynb <==
## Creacion de DataFrame para funciones

vamos a crear los DataFrame de datos auxiliares que serán necesarios para las funciones que vamos a desarrollar.

### Cargamos las librerias necesarias

import pandas as pd
import pyarrow as pa
import pyarrow.parquet as pq
import warnings
import numpy as np
warnings.filterwarnings("ignore")
from fastapi.encoders import jsonable_encoder

### Cargamos los CSV necesarios para extraer la informacion necesaria

df_output = pd.read_csv("data_repo/output_steam_games_limpio.csv", encoding="utf-8")
df_reviews = pd.read_csv("data_repo/user_reviews_limpio.csv", encoding="utf-8")
df_items = pd.read_csv("data_repo/users_items_limpio.csv", encoding="utf-8")

df_output.columns

df_reviews.columns

df_items.columns

Se recopila la informacion de las CSV:
dataframe output (Informacion de juegos)
* app_name : Nombre del juego 
* id : Id de juego
* developer : desarrollador
* price_numeric : Precio de lanzamiento
* release_year : Año de lanzamiento
* main_genre : Genero principal

dataframe Reviews (Reseñas de juegos)
* user_id : Id de usuario
* item_id : Id de juego
* recommend : Si recomienda el juego
* year_posted : Año de recomendacion
* sentiment_analysis : Valorizacion ( 0-Negativo, 1-Neutral, 2-Positivo)

dataframe Items (Informacion de juegos por usuario)
* user_id : Id de usuario
* items_count : Cantidad de juegos por usuario
* item_id : Id del juego
* item_name : Nombre de juego
* playtime_forever : Tiempo jugado en minutos
* playtime_hour : Timpo juego en horas

### Funcion UserForGenre
Debe devolver el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año de lanzamiento

# se extrae columnas relevantes de df_output
userForGenre_output_subset = df_output[["id", "main_genre", "release_year"]]

# se extrae columnas relevantes de df_reviews
userForGenre_reviews_subset = df_reviews[["user_id", "item_id", "year_posted"]]

# se extrae columnas relevantes de df_items
userForGenre_items_subset = df_items[["user_id", "item_id", "playtime_hour"]]

# Se combina en un primer los dataframes
userForGenre_merged = userForGenre_items_subset.merge(userForGenre_reviews_subset, on=["user_id", "item_id"], how="inner")
userForGenre_merged

# Se combina todos en un solo dataframes
userForGenre_merged_2 = userForGenre_merged.merge(userForGenre_output_subset, left_on="item_id", right_on="id", how="inner")
userForGenre_merged_2

# Agrupar por usuario, género principal y año de publicación, sumando las horas jugadas
df_userForGenre = userForGenre_merged_2.groupby(["user_id","main_genre","year_posted"])["playtime_hour"].sum().reset_index()
df_userForGenre

dataframe userForGenre:
* user_id : id de usuario
* main_genre : genero principal
* year_posted : año jugado
* playtime_hour : cantidad de horas juegadas

Guardamos el DataFrame como un CSV y un archivo PARQUET

userForGenre_limpio = "data_repo/df_UsersForGenre.csv"
df_userForGenre.to_csv(userForGenre_limpio, index=False, encoding="utf-8")
print(f"Se guardó el archivo {userForGenre_limpio}") 

userForGenre_limpio_2 = "Datasets/df_UsersForGenre.parquet"
df_userForGenre.to_parquet(userForGenre_limpio_2, index=False)
print(f"Se guardó el archivo {userForGenre_limpio_2}")

## Funcion BestDeveloper
Devuelve el top 3 de desarrolladores con juegos MÁS recomendados por usuarios para el año dado. (reviews.recommend = True y comentarios positivos)

# Se selecciona las columnas 'id' y 'developer' del DataFrame df_output y eliminar las filas duplicadas basadas en 'id'
df_output4 = df_output[["id", "developer"]].drop_duplicates()

# Se selecciona las columnas relevantes para las revisiones: 'item_id', 'recommend', 'year_posted' y 'sentiment_analysis'
df_reviews4 = df_reviews[["item_id", "recommend", "year_posted","sentiment_analysis"]]

# Se fusiona los DataFrames df_reviews4 y df_output4 en función de la coincidencia entre 'item_id' y 'id'
df_reviews_merged4 = pd.merge(df_reviews4, df_output4, left_on="item_id", right_on="id")

# Se filtra las revisiones fusionadas para seleccionar solo aquellas donde la recomendación es Falsa (False) y el análisis de sentimiento es igual a 0
negative_reviews4 = df_reviews_merged4[(df_reviews_merged4["recommend"]== False) & (df_reviews_merged4["sentiment_analysis"] == 0)]

# Se agrupa las revisiones negativas filtradas por año, item_id y desarrollador, contando el número de recomendaciones en cada grupo
grouped_reviews4 = negative_reviews4.groupby(["year_posted", "item_id", "developer"]).size().reset_index(name="recommendation_count")

# Se obtiene las tres filas con las mejores recomendaciones para cada año
df_bestDeveloper = (
    grouped_reviews4.sort_values(by=["year_posted", "recommendation_count"], ascending=[True, False])  # Cambiado a False para orden descendente
                   .groupby("year_posted")
                   .head(3)
                   .reset_index(drop=True)
)

df_bestDeveloper


dataframe bestDeveloper:
* year_posted : año jugado
* item_id : id de juego
* developer	: desarrollador
* recommendation_count : cantidad de recomendaciones

Guardamos el DataFrame como un CSV y un archivo PARQUET

bestDeveloper_limpio = "data_repo/df_BestDeveloper.csv"
df_bestDeveloper.to_csv(bestDeveloper_limpio, index=False, encoding="utf-8")
print(f"Se guardó el archivo {bestDeveloper_limpio}") 

bestDeveloper_limpio_2 = "Datasets/df_BestDeveloper.parquet"
df_bestDeveloper.to_parquet(bestDeveloper_limpio_2, index=False)
print(f"Se guardó el archivo {bestDeveloper_limpio_2}")

## Funcion DeveloperReviews

Según el desarrollador, se devuelve un diccionario con el nombre del desarrollador como llave y una lista con la cantidad total de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento como valor positivo o negativo.

# Se seleciona las columnas relevantes de df_reviews
developerReviews_reviews = df_reviews[["user_id", "item_id","recommend","year_posted","sentiment_analysis"]]
developerReviews_reviews

# Se seleciona las columnas relevantes de df_output
developerReviews_output = df_output[["id","app_name","developer"]]
developerReviews_output

# Se combina los dataframes en base al "items id" y "ID" del juego
developerReviews_merge = pd.merge(developerReviews_reviews, developerReviews_output, left_on="item_id", right_on="id")
developerReviews_merge

# Seleccionamos las columnas deseadas y agrupamos por desarrollador y análisis de sentimiento
developerReviews_merge_2 = developerReviews_merge[["user_id","app_name","developer","recommend","year_posted","sentiment_analysis"]]
df_developerReviews = developerReviews_merge_2.groupby(["developer","sentiment_analysis"]).size().reset_index(name="sentiment_analysis_count")
df_developerReviews

dataframe developerReviews:
* developer	: desarrollador
* sentiment_analysis : tipo de valorizacion
* sentiment_analysis_count : cantidad de valorizacion

Guardamos el DataFrame como un CSV y un archivo PARQUET

developerReviews_limpio = "data_repo/df_DeveloperReviews.csv"
df_developerReviews.to_csv(developerReviews_limpio, index=False, encoding="utf-8")
print(f"Se guardó el archivo {developerReviews_limpio}") 

developerReviews_limpio_2 = "Datasets/df_DeveloperReviews.parquet"
df_developerReviews.to_parquet(developerReviews_limpio_2, index=False)
print(f"Se guardó el archivo {developerReviews_limpio_2}")

## Funcion Developer

Cantidad de items y porcentaje de contenido Free por año según empresa desarrolladora

# Se seleciona las columnas relevantes de df_output 
developer_output_subset = df_output[["id", "developer","price_numeric","release_year"]]
developer_output_subset

# Se agrupa por release_year y developer.
grupo_por_year_dev = developer_output_subset.groupby(['release_year', 'developer'])

# Se cuenta el número de id en cada grupo y crear una nueva columna
cantidad_ids = grupo_por_year_dev['id'].count().reset_index(name='cantidad_id')

# Se calcula el porcentaje de price_numeric igual a 0 para cada grupo y crear una nueva columna
porcentaje_price_numeric_0 = grupo_por_year_dev.apply(lambda x: (x['price_numeric'] == 0).mean() * 100).reset_index(name='contenido free')

# Se combina los resultados en un solo DataFrame
df_developer= cantidad_ids.merge(porcentaje_price_numeric_0, on=['release_year', 'developer'])
df_developer



dataframe Developer:
* release_year: año de lanzamiento
* developer	: desarrollador
* cantidad_id : cantidad de items
* contenido free : % de contenido free por año

Guardamos el DataFrame como un CSV y un archivo PARQUET

developer_limpio = "data_repo/df_Developer.csv"
df_developer.to_csv(developer_limpio, index=False, encoding="utf-8")
print(f"Se guardó el archivo {developer_limpio}") 

developer_limpio_2 = "Datasets/df_Developer.parquet"
df_developer.to_parquet(developer_limpio_2, index=False)
print(f"Se guardó el archivo {developer_limpio_2}")

## Funcion UserData

Debe devolver cantidad de dinero gastado por el usuario, el porcentaje de recomendación en base a reviews.recommend y cantidad de ítems

# se extrae columnas relevantes de df_output
userData_output_subset = df_output[["id", "price_numeric"]]
userData_output_subset


# se extrae columnas relevantes de df_reviews
userData_reviews_subset = df_reviews[["user_id", "item_id", "recommend"]]
userData_reviews_subset

# se extrae columnas relevantes de df_items
userData_items_subset = df_items[["user_id", "item_id", "items_count"]]
userData_items_subset

userData_merged = userData_items_subset.merge(userData_output_subset, left_on="item_id", right_on="id", how="inner")
userData_merged

# Se crea un df auxiliar 
df_gasto_items_usuario = userData_merged.groupby(['user_id','items_count'])['price_numeric'].sum().reset_index()
df_gasto_items_usuario

df_recommend = userData_reviews_subset.groupby('user_id')['recommend'].sum().reset_index()
df_recommend

userData_merged_2 = df_gasto_items_usuario.merge(df_recommend, on = 'user_id', how="inner")
userData_merged_2

userData_merged_2.info()

dataframe UserData:
* items_count : cantidad de items
* user_id : id de usuario
* price_numeric	: dinero gastado
* recommend	: cantidad de recomendaciones

Guardamos el DataFrame como un CSV y un archivo PARQUET

userData_limpio = "data_repo/df_UserData.csv"
userData_merged_2.to_csv(userData_limpio, index=False, encoding="utf-8")
print(f"Se guardó el archivo {userData_limpio}") 

userData_limpio_2 = "Datasets/df_UserData.parquet"
userData_merged_2.to_parquet(userData_limpio_2, index=False)
print(f"Se guardó el archivo {userData_limpio_2}")
