==> 4_API_FUNCIONES.ipynb <==
## Creacion de Funciones para API

Vamos a desarrollar y evaluar las funciones clave de nuestra API utilizando los datos disponibles. Este proceso nos permite confirmar la efectividad y eficiencia de las funciones en un entorno controlado. Sin embargo, al implementarlas en FastAPI y Render, es posible que necesitemos ajustes adicionales para cumplir con sus requisitos específicos

import pandas as pd
import warnings
warnings.filterwarnings("ignore")

## Funcion UserForGenre

df_UserForGenre = pd.read_csv("data_repo\df_UsersForGenre.csv")
df_UserForGenre

def UserForGenre( genero : str ):
    """
    Ingrese genero y devuelve el usuario on más horas jugadas  
    y una lista de la acumulación de horas jugadas por año.
    """
    genero_1 = df_UserForGenre[df_UserForGenre["main_genre"]== genero]
    user_max = genero_1.loc[genero_1["playtime_hour"].idxmax()]["user_id"]
    horas_x_año = genero_1.groupby(["year_posted"])["playtime_hour"].sum().reset_index()
    horas_lista = horas_x_año.rename(columns={'year_posted': 'Año', 'playtime_hour': 'Horas'}).to_dict(orient="records")
    result_ug = {
        f"Usuario con más horas jugadas para Género: {genero}": user_max,
        "Horas jugadas": horas_lista
    }
    return result_ug

UserForGenre("Adventure")

## Funcion BestDeveloper

df_BestDeveloper = pd.read_csv("data_repo\df_BestDeveloper.csv")
df_BestDeveloper

def best_developer_year(año: int):
    """
    Ingrese año y devuelve el top 3 de desarrolladores con juegos MAS 
    recomendados por usuarios.
    """
    if type(año) != int:
        return {"Debes colocar el año en entero, Ejemplo:2012"}
    if año < df_BestDeveloper["year_posted"].min() or año > df_BestDeveloper["year_posted"].max():
        return {"Año no encontrado "}
    df_año = df_BestDeveloper[df_BestDeveloper["year_posted"] == año]
    df_ordenado_recomendacion = df_año.sort_values(by="recommendation_count", ascending=False)
    top_3_developers = df_ordenado_recomendacion.head(3)[["developer", "recommendation_count"]]
    
    # Crear el resultado en el formato deseado
    result_bd = [{"Puesto {}: ".format(i+1) + top_3_developers.iloc[i]['developer'] for i in range(len(top_3_developers))}]
    
    return result_bd



best_developer_year(2014)


## Funcion DeveloperReviews

df_DeveloperReviews = pd.read_csv("data_repo/df_DeveloperReviews.csv")
df_DeveloperReviews

def developer_reviews_analysis(desarrolladora: str):
    """
    Ingrese desarrolladora y devuelve un diccionario con el nombre de la desarrolladora como llave y una lista 
    con la cantidad total de registros de reseñas de usuarios que se encuentren categorizados con 
    un análisis de sentimiento como valor.
    """
    if type(desarrolladora) != str:
        return "Debes colocar un developer de tipo str, EJ:'07th Expansion'"
    if len(desarrolladora) == 0:
        return "Debes colocar un developer en tipo String"
    
    # Se filtra el DataFrame por la desarrolladora dada
    df_developer = df_DeveloperReviews[df_DeveloperReviews["developer"] == desarrolladora]
    
    # Se suma las cuentas de análisis de sentimiento por categoría
    sentiment_counts = df_developer.groupby("sentiment_analysis")["sentiment_analysis_count"].sum().to_dict()
    
    # Se mapea las categorías de análisis de sentimiento a nombres más descriptivos, excluyendo "Neutral"
    sentiment_dicc = {0: "Negativo", 2: "Positivo"}
    
    # Se crea el diccionario resultante
    result_dr = {desarrolladora: [f"{sentiment_dicc[key]} = {value}" for key, value in sentiment_counts.items() if key in sentiment_dicc]}
    
    return result_dr

developer_reviews_analysis('Valve')

## Funcion Developer

df_Developer = pd.read_csv("data_repo/df_Developer.csv")
df_Developer

def developer(desarrollador: str):
    """
    Ingrese desarrollador y devuelve un Diccionario con años como clave y Cantidad de Items y Porcentaje de Contenido Free
    """
    if type(desarrollador) != str:
        return {"Error": "Debes colocar un desarrollador de tipo str, por ejemplo: 'Valve'"}
    if len(desarrollador) == 0:
        return {"Error": "Debes colocar un desarrollador en tipo String"}
    
    df_desarrollador = df_Developer[df_Developer["developer"] == desarrollador]
    df_desarrollador_subset = df_desarrollador[["release_year", "cantidad_id", "contenido free"]]
    df_desarrollador_subset["contenido_free_porcentaje"] = df_desarrollador_subset["contenido free"].round(2)
    result_dict = {}
    for year, data in df_desarrollador_subset.groupby("release_year"):
        result_dict[year] = {str(year): {'Cantidad de Items': cantidad, 'Contenido Free': porcentaje} for cantidad, porcentaje in data[["cantidad_id", "contenido_free_porcentaje"]].values}
    reformatted_result = {year: values[str(year)] for year, values in result_dict.items()}
    return reformatted_result


developer("Valve")


## Funcion UserData

df_UserData = pd.read_csv("data_repo/df_UserData.csv")
df_UserData

def userdata(user_id: str):
    """
    Ingrese id de usuario y devuelve la cantidad de dinero gastado, el porcentaje de 
    recomendación y cantidad de juegos.
    """
    user_data = df_UserData[df_UserData['user_id'] == user_id]
    
    if user_data.empty:
        return {"Error": f"No se encontraron datos para el user_id '{user_id}'"}
    
    price_numeric_total = user_data.iloc[0]['price_numeric']
    recommend_percentage = (user_data['recommend'].sum() / len(user_data)) * 100
    items_count = user_data.iloc[0]['items_count']
    result = {
        "user_id": user_id,
        "price_numeric": f"{price_numeric_total} USD",
        "% de recomendación": f"{recommend_percentage}%",
        "items_count": items_count
    }
    
    return result

userdata('zynxgameth')
